[fix] 라우터 구조 변경
