+++ src/sgs-client/src/components/LoginForm.vue
@@ -113,6 +113,7 @@ export default {
         await this.LOGIN(userData);
         this.$router.push("/channels/@me");
       } catch (err) {
+        console.log("로그인 실패에러", err.response);
         this.logMessage = "로그인에 실패하셨습니다.";
       }
     },
