+++ src/sgs-client/src/pages/MainPage.vue
@@ -1,9 +1,159 @@
 <template>
-  <div></div>
+  <div>
+    <div class="wrapper2">
+      <div class="wrapper">
+        <div class="container">
+          <navigation-bar v-if="navbar"></navigation-bar>
+          <router-view></router-view>
+        </div>
+      </div>
+    </div>
+    <server-popout></server-popout>
+    <friends-plus-action></friends-plus-action>
+    <create-server-modal></create-server-modal>
+    <create-channel-modal></create-channel-modal>
+    <friends-delete-modal></friends-delete-modal>
+    <friends-block-modal></friends-block-modal>
+    <friends-profile-modal></friends-profile-modal>
+    <create-category-modal></create-category-modal>
+    <category-setting-modal></category-setting-modal>
+    <category-delete-modal></category-delete-modal>
+    <server-setting-modal></server-setting-modal>
+  </div>
 </template>
 
 <script>
-export default {};
+import Stomp from "webstomp-client";
+import SockJS from "sockjs-client";
+import { mapGetters, mapMutations, mapState } from "vuex";
+import { getBaseURL } from "../utils/common";
+import NavigationBar from "../components/NavigationBar.vue";
+import CreateServerModal from "../components/CreateServerModal.vue";
+import CreateChannelModal from "../components/CreateChannelModal.vue";
+import FriendsPlusAction from "../components/common/FriendsPlusAction.vue";
+import FriendsDeleteModal from "../components/FriendsDeleteModal.vue";
+import ServerPopout from "../components/ServerPopout.vue";
+import FriendsBlockModal from "../components/FriendsBlockModal.vue";
+import FriendsProfileModal from "../components/FriendsProfileModal.vue";
+import CreateCategoryModal from "../components/CreateCategoryModal.vue";
+import CategorySettingModal from "../components/CategorySettingModal.vue";
+import CategoryDeleteModal from "../components/CategoryDeleteModal.vue";
+import ServerSettingModal from "../components/ServerSettingModal.vue";
+
+export default {
+  name: "App",
+  components: {
+    NavigationBar,
+    CreateServerModal,
+    CreateChannelModal,
+    FriendsPlusAction,
+    FriendsDeleteModal,
+    ServerPopout,
+    FriendsBlockModal,
+    FriendsProfileModal,
+    CreateCategoryModal,
+    CategorySettingModal,
+    CategoryDeleteModal,
+    ServerSettingModal,
+  },
+  data() {
+    return {
+      navbar: true,
+    };
+  },
+  created() {
+    console.log("로그인안된 크레이트");
+    if (this.getEmail) {
+      console.log("안녕크레이트");
+      this.connect();
+    }
+    const url = window.location.pathname;
+    if (url == "/settings") {
+      this.navbar = false;
+      console.log("navbar", this.navbar);
+    } else {
+      this.navbar = true;
+      console.log("navbar", this.navbar);
+    }
+  },
+  watch: {
+    // 라우터의 변경을 감시
+    $route(to, from) {
+      if (to.path != from.path) {
+        if (to.path === "/settings") {
+          this.navbar = false;
+          console.log("navbar", this.navbar);
+        } else {
+          this.navbar = true;
+          console.log("navbar", this.navbar);
+        }
+      }
+    },
+  },
+  computed: {
+    ...mapState("utils", ["stompSocketClient", "stompSocketConnected"]),
+    ...mapGetters("user", ["getEmail", "getUserId", "getAccessToken"]),
+  },
+  methods: {
+    ...mapMutations("utils", [
+      "setStompSocketClient",
+      "setStompSocketConnected",
+    ]),
+    connect() {
+      const serverURL = getBaseURL() + "my-chat";
+      let socket = new SockJS(serverURL);
+      this.setStompSocketClient(Stomp.over(socket));
+      console.log(`소켓 연결을 시도합니다. 서버 주소: ${serverURL}`);
+      this.stompSocketClient.connect(
+        {
+          "access-token": this.getAccessToken,
+          "user-id": this.getUserId,
+        }, //header
+        (frame) => {
+          // 소켓 연결 성공
+          this.connected = true;
+          this.setStompSocketConnected(true);
+          console.log("소켓 연결 성공", frame);
+        },
+        (error) => {
+          // 소켓 연결 실패
+          console.log("소켓 연결 실패", error);
+          this.connected = false;
+          this.setStompSocketConnected(false);
+        }
+      );
+    },
+  },
+};
 </script>
 
-<style></style>
+<style>
+@import "../css/common.css";
+.wrapper2 {
+  position: absolute;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  background-color: var(--dark-grey-color);
+  display: flex;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+  flex-direction: column;
+}
+.container {
+  position: relative;
+  overflow: hidden;
+  width: 100%;
+  height: 100%;
+  display: flex;
+}
+button {
+  font-weight: 500;
+  border: 0;
+  cursor: pointer;
+}
+a img {
+  border: none;
+}
+</style>
