+++ src/sgs-client/src/pages/NotFoundPage.vue
@@ -6,5 +6,4 @@
 export default {};
 </script>
 
-<style>
-</style>
+<style></style>
