+++ src/sgs-client/src/routes/index.js
@@ -8,75 +8,61 @@ import UserSettingPage from "../pages/UserSettingPage.vue";
 import ServerPage from "../pages/ServerPage.vue";
 import ServerWelcomePage from "../pages/ServerWelcomePage.vue";
 import PrivateDMPage from "../pages/PrivateDMPage.vue";
+import MainPage from "../pages/MainPage.vue";
 
-import store from "../store/index";
 Vue.use(VueRouter);
 
 export const router = new VueRouter({
   mode: "history",
   routes: [
     {
-      path: "/",
-      redirect: () => {
-        const isLogin = store.getters["user/getEmail"];
-        if (!isLogin) {
-          return "/login";
-        } else {
-          return "/channels/@me";
-        }
-      },
-    },
-    {
       path: "/login",
-      name: "login",
+      name: "LoginPage",
       component: LoginPage,
     },
     {
       path: "/register",
-      name: "registerPage",
+      name: "RegisterPage",
       component: RegisterPage,
     },
     {
-      path: "/channels/@me/:id",
-      name: "PrivateDMPage",
-      component: PrivateDMPage,
-      meta: { auth: true },
-    },
-    {
-      path: "/channels/@me",
-      name: "MyPage",
-      component: MyPage,
-      meta: { auth: true },
-    },
-    {
-      path: "/settings",
-      name: "UserSettingPage",
-      component: UserSettingPage,
-      meta: { auth: true },
-    },
-    {
-      path: "/channels/:serverid/:channelid",
-      name: "ServerPage",
-      component: ServerPage,
-    },
-    {
-      path: "/channels/:serverid",
-      name: "ServerWelcomePage",
-      component: ServerWelcomePage,
+      path: "/",
+      name: "MainPage",
+      component: MainPage,
+      children: [
+        {
+          path: "channels/@me/:id",
+          name: "privateDmPage",
+          component: PrivateDMPage,
+          meta: { auth: true },
+        },
+        {
+          path: "channels/@me",
+          name: "MyPage",
+          component: MyPage,
+          meta: { auth: true },
+        },
+        {
+          path: "settings",
+          name: "UserSettingPage",
+          component: UserSettingPage,
+          meta: { auth: true },
+        },
+        {
+          path: "channels/:serverid/:channelid",
+          name: "ServerPage",
+          component: ServerPage,
+        },
+        {
+          path: "channels/:serverid",
+          name: "ServerWelcomePage",
+          component: ServerWelcomePage,
+        },
+      ],
     },
     {
       path: "*",
       component: NotFoundPage,
     },
   ],
-});
-
-router.beforeEach((to, from, next) => {
-  const isLogin = store.getters["user/getEmail"];
-  if (to.meta.auth && !isLogin) {
-    alert("인증이 필요합니다.");
-    next("/login");
-    return;
-  }
-  next();
 });
